refactor: 주먹 검출을 방해하는 엄지 조건 제거
