+++ src/pages/lantern/Lantern.tsx
@@ -3,9 +3,9 @@ import { useNavigate, useSearchParams } from 'react-router-dom';
 import * as styles from './Lantern.css';
 import { generateNonOverlappingPositions, seededRandom } from './utils';
 import { CloseButton } from '../lanternDetail/components/CloseButton/CloseButton';
-import { useCloseGesture } from '../lanternDetail/hooks/useCloseGesture';
-import { useHandMark } from '../../components/common/lantern/hooks/useHandMark';
 import { VideoFeed } from '../../components/common/lantern/VideoFeed';
+import { useCloseGesture } from '../../hooks/useCloseGesture';
+import { useHandMark } from '../../hooks/useHandMark';
 import { LanternListResponse } from '@/apis/lantern';
 import { MusicStatusData } from '@/apis/lantern/subscribeStatus';
 import hand from '@/assets/hand.png';
@@ -13,7 +13,7 @@ import LanternImg1 from '@/assets/Lantern.svg?react';
 import LanternImg2 from '@/assets/RoundLantern.svg?react';
 import { Alert } from '@/components/common/alert';
 import { LanternWithRect } from '@/components/common/lantern/constants';
-import { useLanternHit } from '@/components/common/lantern/hooks/useLanternHit';
+import { useLanternHit } from '@/hooks/useLanternHit';
 import { lanternsDetail, lanternsList } from '@/mocks';
 // import { useLanternList } from '@/queries/lantern/getLanternList';
 import { useCachedLanternProgress } from '@/queries/lantern/useLanternProgress';
