+++ src/pages/lanternDetail/LanternDetail.tsx
@@ -4,13 +4,13 @@ import AudioVisualizer from './components/AudioVisualizer/AudioVisualizer';
 import { CloseButton } from './components/CloseButton/CloseButton';
 import { useAudioPlayer } from './hooks/useAudioPlayer';
 import { useCarousel } from './hooks/useCarousel';
-import { useCloseGesture } from './hooks/useCloseGesture';
-// import { useLanternDetail } from './hooks/useLanternDetail';
 import * as styles from './LanternDetail.css';
+import { useCloseGesture } from '../../hooks/useCloseGesture';
+// import { useLanternDetail } from './hooks/useLanternDetail';
 import hand from '@/assets/hand.png';
-import { useHandMark } from '@/components/common/lantern/hooks/useHandMark';
 import { VideoFeed } from '@/components/common/lantern/VideoFeed';
 import { Toast } from '@/components/common/toast';
+import { useHandMark } from '@/hooks/useHandMark';
 import { lanternType } from '@/mocks';
 import { queryClient } from '@/queries/queryClient';
 import { lanternKeys } from '@/queries/queryKey';
