+++ src/utils/index.ts
@@ -1,5 +1,5 @@
 import { Keypoint } from '@tensorflow-models/hand-pose-detection';
-import { Rect } from '../constants';
+import { Rect } from '../components/common/lantern/constants';
 
 // 풍등 내부 여부 함수
 export const isInside = (pos: { x: number; y: number }, rect: Rect) => {
@@ -7,20 +7,11 @@ export const isInside = (pos: { x: number; y: number }, rect: Rect) => {
 };
 
 // 주먹 확인
-export const isFist = (marks: Keypoint[], handedness: 'Left' | 'Right') => {
-  // 엄지 검출
-  let thumbCrossed;
-  if (handedness === 'Right') {
-    thumbCrossed = marks[4].x > marks[8].x;
-  } else {
-    thumbCrossed = marks[4].x < marks[8].x;
-  }
-
+export const isFist = (marks: Keypoint[]) => {
   return (
     marks[8].y > marks[6].y && // 검지 접힘
     marks[12].y > marks[10].y && // 중지 접힘
     marks[16].y > marks[14].y && // 약지 접힘
-    marks[20].y > marks[18].y && // 소지 접힘
-    thumbCrossed
+    marks[20].y > marks[18].y // 소지 접힘
   );
 };
